Break z-axis step out into object
